pyproject.toml
[project]
name = "lit-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "uvicorn[standard]>=0.30.0",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
    "openai>=2.5.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.0",
]
